requirements.txt
streamlit==1.26.0
opencv-python-headless==4.8.1.78
pillow==9.5.0
numpy==1.26.0
